Create a base class for dict tree classes
This generalizes some of the functions and moves them into a class
hierarchy. We intend to add a new class for multiple columns to be
used along with the base dict tree classes.
